# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "Jungle Gaming Financial API"
  version: "0.12.0"
  description: "Fase 12. JSON financeiro estrito; body máximo 64 KiB. Money sem float. Bearer Keycloak via client_credentials. Wallet/ledger/reconciliation/metrics internal-only; transações provider-scoped. Health público. Reconciliation read-only em snapshot consistente. Métricas agregadas Prometheus sem identificadores financeiros nas labels."
servers:
  -
    url: "http://localhost:8080"
    description: "Desenvolvimento local"
security:
  -
    BearerAuth: []
paths:
...
  "/wallets":
    post:
      operationId: "createWallet"
      description: "Somente internal. Sem Idempotency-Key para criação; (playerId,currency) é único."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/CreateWalletRequest"
      responses:
        "201":
          description: "Operação criada; Location aponta para o recurso."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/CreateWalletResult"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "409":
          "$ref": "#/components/responses/Conflict"
        "413":
          "$ref": "#/components/responses/PayloadTooLarge"
        "415":
          "$ref": "#/components/responses/UnsupportedMediaType"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
  "/wallets/{walletId}":
    get:
      operationId: "getWallet"
      description: "Somente internal; não existe ownership provider-wallet seguro no modelo."
      parameters:
        -
          name: "walletId"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/Identifier"
      responses:
        "200":
          description: "Recurso consultado."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Wallet"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
  "/wallets/{walletId}/ledger":
    get:
      operationId: "getWalletLedger"
      description: "Somente internal. Ordem crescente (walletVersion, ledgerEntryId com collation C). Keyset, sem offset; cursor não é grant de acesso nem snapshot fixo. Novas entradas podem aparecer nas páginas seguintes."
      parameters:
        -
          name: "walletId"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/Identifier"
        -
          name: "limit"
          in: "query"
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
            default: 50
        -
          name: "cursor"
          in: "query"
          schema:
            type: "string"
            minLength: 1
            maxLength: 2048
          description: "Cursor opaco retornado; wallet-bound. Inválido/vazio ou parâmetros duplicados resultam 400."
      responses:
        "200":
          description: "Recurso consultado."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/LedgerPage"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
  "/wagering/transactions":
    post:
      operationId: "processWager"
      description: "Provider próprio; internal não pode impersonar. Resultados de negócio são persistidos. Replay usa observedBalance original, sem reler saldo atual. Referência pendente resolvida pode mudar estado/replay no worker."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/WagerRequest"
      responses:
        "200":
          description: "Replay PROCESSED; idempotentReplay=true, observedBalance original."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WagerResult"
        "201":
          description: "Operação criada; Location aponta para o recurso."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WagerResult"
        "202":
          description: "PENDING_REFERENCE persistido; replay também 202 enquanto pendente."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WagerResult"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "409":
          "$ref": "#/components/responses/Conflict"
        "413":
          "$ref": "#/components/responses/PayloadTooLarge"
        "415":
          "$ref": "#/components/responses/UnsupportedMediaType"
        "422":
          description: "REJECTED/FAILED persistido; failureCode estável, inclusive no replay. Não é falha de infraestrutura."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/WagerResult"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
      parameters:
        -
          name: "Idempotency-Key"
          in: "header"
          required: true
          schema:
            "$ref": "#/components/schemas/Identifier"
          description: "Uma única key opaca, obrigatória. Namespace persistente por provider; não gerada pelo servidor. Mesma key/payload replays; payload diferente 409."
  "/wagering/transactions/{transactionId}":
    get:
      operationId: "getTransaction"
      description: "Own provider ou internal. Foreign existente é 403; ausente é 404. Autorização precede resposta."
      parameters:
        -
          name: "transactionId"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/Identifier"
      responses:
        "200":
          description: "Recurso consultado."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Transaction"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
  "/providers/{providerId}/wagering/transactions/{externalTransactionId}":
    get:
      operationId: "getExternalTransaction"
      description: "Own provider ou internal. ProviderId na URL é comparado com principal antes da query."
      parameters:
        -
          name: "providerId"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/Identifier"
        -
          name: "externalTransactionId"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/Identifier"
      responses:
        "200":
          description: "Recurso consultado."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Transaction"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
  "/wallets/{walletId}/reconciliation":
    post:
      operationId: "reconcileWallet"
      description: "Bearer internal-service com role internal. Auditoria read-only da wallet, ledger completo e transações relacionadas em REPEATABLE READ READ ONLY. Sem body ou Idempotency-Key obrigatório. Não altera dados, não cria Inbox/Outbox e não corrige divergências. CONSISTENT e DIVERGENT retornam 200. checkedAt é o instante UTC de conclusão, não um token de snapshot."
      parameters:
        -
          name: "walletId"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/Identifier"
      responses:
        "200":
          description: "Auditoria concluída: CONSISTENT sem divergências ou DIVERGENT com evidências estruturadas."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/ReconciliationResult"
              examples:
                consistent:
                  value:
                    walletId: "wallet-example"
                    currency: "BRL"
                    status: "CONSISTENT"
                    storedBalance: {amount: "100.00", currency: "BRL"}
                    calculatedBalance: {amount: "100.00", currency: "BRL"}
                    difference: {amount: "0.00", currency: "BRL"}
                    consistent: true
                    checkedEntries: 1
                    walletBalance: {amount: "100.00", currency: "BRL"}
                    ledgerBalance: {amount: "100.00", currency: "BRL"}
                    walletVersion: 1
                    expectedWalletVersion: 1
                    ledgerEntriesChecked: 1
                    transactionsChecked: 1
                    divergences: []
                    checkedAt: "2026-10-07T12:00:00Z"
                divergent:
                  value:
                    walletId: "wallet-example"
                    currency: "BRL"
                    status: "DIVERGENT"
                    storedBalance: {amount: "99.00", currency: "BRL"}
                    calculatedBalance: {amount: "100.00", currency: "BRL"}
                    difference: {amount: "-1.00", currency: "BRL"}
                    consistent: false
                    checkedEntries: 1
                    walletBalance: {amount: "99.00", currency: "BRL"}
                    ledgerBalance: {amount: "100.00", currency: "BRL"}
                    walletVersion: 1
                    expectedWalletVersion: 1
                    ledgerEntriesChecked: 1
                    transactionsChecked: 1
                    divergences:
                      - code: "BALANCE_MISMATCH"
                        expected: "100.00"
                        actual: "99.00"
                        message: "wallet balance differs from reconstructed ledger balance"
                    checkedAt: "2026-10-07T12:00:00Z"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
...
  "/auth/me":
    get:
      operationId: "currentPrincipal"
      description: "Probe mantido da fase 8; token válido."
      parameters: []
      responses:
        "200":
          description: "Recurso consultado."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Principal"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
  "/auth/internal":
    get:
      operationId: "internalPrincipal"
      description: "Probe mantido da fase 8; role internal."
      parameters: []
      responses:
        "200":
          description: "Recurso consultado."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Principal"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
  "/auth/providers/{providerId}":
    get:
      operationId: "providerPrincipal"
      description: "Probe mantido da fase 8; own provider ou internal."
      parameters:
        -
          name: "providerId"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/Identifier"
      responses:
        "200":
          description: "Recurso consultado."
          headers:
            X-Correlation-ID:
              schema:
                type: "string"
              description: "ID propagado/gerado pelo middleware."
            X-Content-Type-Options:
              schema:
                type: "string"
                enum:
                  - "nosniff"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Principal"
        "400":
          "$ref": "#/components/responses/BadRequest"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "500":
          "$ref": "#/components/responses/Internal"
        "503":
          "$ref": "#/components/responses/Unavailable"
components:
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "RS256 access token do realm jungle-gaming, audience wagering-api. Roles provider/internal e provider_id assinado."
  schemas:
...
    Identifier:
      type: "string"
      minLength: 1
      maxLength: 256
      description: "Identificador textual de 1–256 bytes UTF-8; sem whitespace, controles, /, \\, ? ou #. IDs gerados são texto hexadecimal, não UUID."
    Money:
      type: "object"
      additionalProperties: false
      required:
        - "amount"
        - "currency"
      properties:
        amount:
          type: "string"
          pattern: "^(0|[1-9][0-9]*)\\.[0-9]{2}$"
          example: "25.00"
          description: "Exatamente duas casas; string não negativa, sem expoente/sinal; máximo 92233720368547758.07 (int64 cents)."
        currency:
          type: "string"
          pattern: "^[A-Z]{3}$"
          enum: [BRL, USD, EUR]
          example: "BRL"
    CreateWalletRequest:
      type: "object"
      additionalProperties: false
      required:
        - "playerId"
      properties:
        playerId:
          "$ref": "#/components/schemas/Identifier"
        currency:
          type: "string"
          pattern: "^[A-Z]{3}$"
          enum: [BRL, USD, EUR]
          example: "BRL"
        initialBalance:
          type: "object"
          nullable: true
          additionalProperties: false
          required: ["amount", "currency"]
          properties:
            amount:
              type: "string"
              pattern: "^(0|[1-9][0-9]*)\\.[0-9]{2}$"
            currency:
              type: "string"
              pattern: "^[A-Z]{3}$"
              enum: [BRL, USD, EUR]
      anyOf:
        - required: [currency]
        - required: [initialBalance]
          properties:
            initialBalance:
              nullable: false
      description: "Somente internal. currency pode ser inferida de initialBalance.currency. Sem initialBalance, informar currency para abertura zero. Quando ambos presentes devem coincidir. Saldo positivo cria OPENING/ledger/outbox atomicamente."
    CreateWalletResult:
      type: "object"
      additionalProperties: false
      required:
        - "walletId"
        - "playerId"
        - "balance"
        - "walletVersion"
      properties:
        id:
          "$ref": "#/components/schemas/Identifier"
        version:
          type: "integer"
          format: "int64"
          minimum: 1
        walletId:
          "$ref": "#/components/schemas/Identifier"
        playerId:
          "$ref": "#/components/schemas/Identifier"
        balance:
          "$ref": "#/components/schemas/Money"
        walletVersion:
          type: "integer"
          format: "int64"
          minimum: 1
        openingTransactionId:
          "$ref": "#/components/schemas/Identifier"
    Wallet:
      type: "object"
      additionalProperties: false
      required:
        - "walletId"
        - "playerId"
        - "currency"
        - "balance"
        - "version"
      properties:
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
        walletId:
          "$ref": "#/components/schemas/Identifier"
        playerId:
          "$ref": "#/components/schemas/Identifier"
        currency:
          type: "string"
          pattern: "^[A-Z]{3}$"
          enum: [BRL, USD, EUR]
          example: "BRL"
        balance:
          "$ref": "#/components/schemas/Money"
        version:
          type: "integer"
          format: "int64"
          minimum: 1
...
    LedgerPage:
      type: "object"
      additionalProperties: false
      required:
        - "items"
      properties:
        items:
          type: "array"
          items:
            "$ref": "#/components/schemas/LedgerEntry"
        nextCursor:
          type: "string"
          description: "Cursor opaco URL-safe; ausente no fim. Continuar com o mesmo walletId."
...
    Principal:
      type: "object"
      additionalProperties: false
      required:
        - "subject"
        - "clientId"
        - "roles"
        - "internal"
      properties:
        subject:
          type: "string"
        clientId:
          type: "string"
        providerId:
          type: "string"
        roles:
          type: "array"
          nullable: true
          items:
            type: "string"
        internal:
          type: "boolean"
  responses:
    BadRequest:
      description: "400: JSON, campos, Money, identificadores, cursor/limit ou Idempotency-Key inválidos."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    Unauthorized:
      description: "401: token ausente/inválido ou autenticação indisponível."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
        WWW-Authenticate:
          schema:
            type: "string"
            enum:
              - "Bearer"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    Forbidden:
      description: "403: identidade válida sem permissão ou provider mismatch."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    NotFound:
      description: "404: recurso inexistente."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    Conflict:
      description: "409: wallet existente, ID externo duplicado ou key com payload diferente."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    PayloadTooLarge:
      description: "413: body maior que 64 KiB."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    UnsupportedMediaType:
      description: "415: exige application/json (charset UTF-8 opcional)."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    Unavailable:
      description: "503: dependência/contexto indisponível; operação não confirmada não é um resultado de negócio."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    Internal:
      description: "500: falha não classificada; sem detalhes internos."
      headers:
        X-Correlation-ID:
          schema:
            type: "string"
          description: "ID propagado/gerado pelo middleware."
        X-Content-Type-Options:
          schema:
            type: "string"
            enum:
              - "nosniff"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
...